Customize GH release notes & add tarball
I wasn't extremely happy with the default github-generated release
notes, specifically the changelog tag detection, so I used the
suggestions I found in the softprops repo: mikepenz's
release-changelog-builder-action.

Added .github/changelog.json to customize the labels, categories, and
output templates for the release notes.

For it to work how I'd like, I had to manually construct the compare
link from the action's detected tags and manually construct the NPM
package url.

Also had to add a manual npm pack step in order to provide the .tgz in
the GitHub release assets, but that can be removed if desired.