Add version.yml workflow for on-click release
Added a new workflow file: version.yml. This workflow simply runs npm
version with the desired values and pushes the commit and tag back to
the repository. Along with release.yml, this functionally creates
on-demand npm and github releases from the click of a button.

This workflow requires a PAT in-place of the GITHUB_TOKEN, as the
workflow default GITHUB_TOKEN will **not** trigger GitHub Actions.