Switch version.yml to hardcoded preid 'beta'
After discussion in the PR, I've decided it is actually better to reduce
the complexity and hardcode the preId when making pre* level versions.

Leaving my comment blocks in this commit for posterity, but I'll delete
them in the next because I don't think it needs to exist in master.