Add addition blockedsource test + data
Created a small blocked COG file in test/data/setup_data.sh (32x32,
BLOCKSIZE=16) and a test that shows the failure (without the fix from
PR#474 applied). The test also confirmed that the failure is (likely)
due to the top value equaling blockSize * blocks.length, as the failure
only occurs with blockSize: 16*16 (as expected).